User repo done
Should be done, also containing GetUsersFromEvent

+++ backend/Repositories/UserRepository.cs
@@ -20,14 +20,34 @@ public class UserRepository
 
     public async Task<User?> GetUserByID(string userId)
     {
+        if (string.IsNullOrEmpty(userId))
+        {
+            throw new ArgumentNullException("User cannot be null or empty.", nameof(userId));
+        }
         return await _context.Users.FindAsync(userId);
     }
 
-    //Searchusers
+    public async Task<ICollection<User>> SearchUsers(string searchString)
+    {
+        int maxresults = 40;
+        IQueryable<User> query = _context.Users;
+
+        if (!string.IsNullOrEmpty(searchString))
+        {
+            query = query.Where(u => (u.Firstname + " " + u.Lastname).Contains(searchString));
+        }
+        query = query.Take(maxresults);
+
+        return await query.ToListAsync();
+    }
 
     public async Task<User?> UpdateUser(User user)
     {
-
+        if (user == null)
+        {
+            // Exception on faulty input
+            throw new ArgumentNullException("User cannot be null or empty.", nameof(user));
+        }
         User? userToChange = await _context.Users.FindAsync(user.UserID);
         if (userToChange != null)
         {
@@ -39,16 +59,35 @@ public class UserRepository
 
     public async Task<User?> DeleteUser(string userId)
     {
+
+        if (string.IsNullOrEmpty(userId))
+        {
+            // Exception on faulty input
+            throw new ArgumentException("User ID cannot be null or empty.", nameof(userId));
+        }
+
         User? userToDelete = await _context.Users.FindAsync(userId);
 
         if (userToDelete != null)
         {
             _context.Users.Remove(userToDelete);
-            await _context.SaveChangesAsync();
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                //log error
+            }
         }
         return userToDelete;
     }
 
-    //getUsersFromEvent
+    public async Task<ICollection<User>> GetUsersFromEvent(int eventId)
+    {
+        return await _context.Users
+        .Where(u => u.EventRelations != null && u.EventRelations.Any(er => er.EventID == eventId))
+        .ToListAsync();
+    }
 
 }
